feat(docat): reorganize module and fix gunicorn config
