+++ docat/tests/test_utils.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from mock import MagicMock, mock_open, patch
 
-from docat.docat.utils import create_nginx_config, create_symlink, extract_archive, remove_docs
+from docat.utils import create_nginx_config, create_symlink, extract_archive, remove_docs
 
 
 def test_symlink_creation():
@@ -60,7 +60,7 @@ def test_create_nginx_config():
     Tests the creation of the nginx config
     """
     jinja_template_mock = "{{ project }}:{{ dir_path }}"
-    with patch("builtins.open", mock_open(read_data=jinja_template_mock)), patch.object(Path, "exists") as mock_exists, patch.object(
+    with patch.object(Path, "read_text", return_value=jinja_template_mock), patch.object(Path, "exists") as mock_exists, patch.object(
         Path, "open", mock_open()
     ) as mock_config, patch("subprocess.run"):
         mock_exists.return_value = False
@@ -90,7 +90,7 @@ def test_not_overwrite_nginx_config():
 def test_archive_artifact():
     target_file = Path("/some/zipfile.zip")
     destination = "/tmp/null"
-    with patch.object(Path, "unlink") as mock_unlink, patch("docat.docat.utils.ZipFile") as mock_zip:
+    with patch.object(Path, "unlink") as mock_unlink, patch("docat.utils.ZipFile") as mock_zip:
         mock_zip_open = MagicMock()
         mock_zip.return_value.__enter__.return_value.extractall = mock_zip_open
 
@@ -103,7 +103,7 @@ def test_archive_artifact():
 
 def test_remove_version(temp_project_version):
     docs, config = temp_project_version("project", "1.0")
-    with patch("docat.docat.utils.UPLOAD_FOLDER", docs), patch("docat.docat.utils.NGINX_CONFIG_PATH", config):
+    with patch("docat.utils.UPLOAD_FOLDER", docs), patch("docat.utils.NGINX_CONFIG_PATH", config):
         remove_docs("project", "1.0")
 
         assert docs.exists()
